+++ CHANGELOG.md
@@ -1,3 +1,35 @@
-## 1.0.0
+## 0.0.1
 
 - Initial version.
+
+## 0.0.2
+
+- Fix parameters
+
+## 0.0.3
+
+- Fix parameters request
+
+## 0.0.4
+
+- Fix parameters request
+
+## 0.0.5
+
+- Fix code
+
+## 0.0.6
+
+- Fix parameter url
+
+## 0.0.7
+
+- Fix parameter url
+
+## 0.0.8
+
+- Fix parameter url
+
+## 0.0.9
+
+- Fix parameter url